lots more updates and added service calls. changed a few floats to doubles for accuracy. fixed corp roles truncation on initial load (from uint32 to uint64) started a random location class with methods to generate a random planet, moon, points between 2 or 3 planets, and a random point within the solar system. more updates to mapData...still not complete. more updates to search....works, but return isnt right. more updates to undock.....still not updating client correctly. updated agents std::vector to std::list for faster iteration. added fleet enums for later began console command system with one command for now.  (exit)

+++ src/eve-common/destiny/DestinyStructs.h
@@ -104,7 +104,7 @@ enum {
 struct BallHeader {
     uint64 entityID;
     uint8 mode;        //see DSTBALL_
-    float radius;    //7.230769230769231e-3
+    double radius;    //7.230769230769231e-3
     double x;    //1.33638303942e12
     double y;    //6.252761088e10
     double z;    //4.1517330432e11
@@ -117,7 +117,7 @@ struct MassSector {
     uint8  cloak;        // indicates cloak
     uint64 allianceID;
     uint32 corpID;
-    float Harmonic;
+    double Harmonic;
 };
 
 //only included if the thing can move...
@@ -126,15 +126,15 @@ struct ShipSector {
     double velocity_x;
     double velocity_y;
     double velocity_z;
-    float agility;
-    float speed_fraction;
+    double agility;
+    double speed_fraction;
 };
 
 struct MiniBall {
     double x;
     double y;
     double z;
-    float radius;
+    double radius;
 };
 
 struct MiniBallList {
